up: update the repo data,dao,entity layers logic

+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dto/AdminConfigDTO.java
@@ -32,16 +32,16 @@ public class AdminConfigDTO {
     /**
      * Delete time
      */
-    private Integer deleteTime;
+    private Long deleteTime;
 
     /**
      * Update time
      */
-    private Integer updateTime;
+    private Long updateTime;
 
     /**
      * Create time
      */
-    private Integer createTime;
+    private Long createTime;
 }
