+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dto/AdminRuleDTO.java
@@ -49,16 +49,16 @@ public class AdminRuleDTO {
     /**
      * Delete time
      */
-    private Integer deleteTime;
+    private Long deleteTime;
 
     /**
      * Update time
      */
-    private Integer updateTime;
+    private Long updateTime;
 
     /**
      * Create time
      */
-    private Integer createTime;
+    private Long createTime;
 }
