+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dto/NotifyContactDTO.java
@@ -42,16 +42,16 @@ public class NotifyContactDTO {
     /**
      * Delete time
      */
-    private Integer deleteTime;
+    private Long deleteTime;
 
     /**
      * Update time
      */
-    private Integer updateTime;
+    private Long updateTime;
 
     /**
      * Create time
      */
-    private Integer createTime;
+    private Long createTime;
 }
