+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dto/NotifyGroupDTO.java
@@ -37,16 +37,16 @@ public class NotifyGroupDTO {
     /**
      * Delete time
      */
-    private Integer deleteTime;
+    private Long deleteTime;
 
     /**
      * Update time
      */
-    private Integer updateTime;
+    private Long updateTime;
 
     /**
      * Create time
      */
-    private Integer createTime;
+    private Long createTime;
 }
